+++ Runtime/GfxDevice/GfxDeviceVulkan.cpp
@@ -10,7 +10,6 @@
 #include <vector>
 #include <optional>
 #include <set>
-
 #include "Sailor.h"
 
 using namespace Sailor;
@@ -106,7 +105,7 @@ void GfxDeviceVulkan::Initialize(const Window* viewport, bool bInIsEnabledValida
 	VkInstanceCreateInfo createInfo{ VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO };
 	createInfo.pApplicationInfo = &appInfo;
 	createInfo.ppEnabledExtensionNames = extensions.data();
-	createInfo.enabledExtensionCount = extensions.size();
+	createInfo.enabledExtensionCount = (uint32_t)extensions.size();
 
 	const std::vector<const char*> validationLayers = { "VK_LAYER_KHRONOS_validation" };
 
@@ -114,7 +113,7 @@ void GfxDeviceVulkan::Initialize(const Window* viewport, bool bInIsEnabledValida
 	if (m_pInstance->bIsEnabledValidationLayers)
 	{
 		createInfo.ppEnabledLayerNames = validationLayers.data();
-		createInfo.enabledLayerCount = validationLayers.size();
+		createInfo.enabledLayerCount = (uint32_t)validationLayers.size();
 
 		PopulateDebugMessengerCreateInfo(debugCreateInfo);
 		createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
